changes to ber.R and READ ME
Changes to ber
1) Changed code to operate from any directory.  Previous code would
throw an error because the working directory was not set properly
2) Found an error that if there is “missing data” in the “Behavior”
category, then the code would not run.  Changed the code to pass a
message saying  that there was missingness in the behavior column and
stop processing the file.  Returns NA information for that file.
3) There is a difference between the “FinnBrain” files and the “PEIDS”
files where instantaneous events are listed as “State point” in the
first and “Point” in the second.  Changed the code to handle both
descriptions.

Changes to README
1) Changed the reference section
2) Code reran